World Quant University project - Computer Vision
This project was one of the projects I embarked on in World Quant University, Computer Vision Track. The Jupyter notebooks belong to World Quant.

In this project, you'll look at a data science competition helping scientists track animals in a wildlife preserve. The goal is to take images from camera traps and classify which animal, if any, is present. To complete the competition, you'll expand your machine learning skills by creating more powerful neural network models that can take images as inputs and classify them into one of multiple categories.

Some of the things you'll learn are:

How to read image files and prepare them for machine learning
How to use PyTorch to manipulate tensors and build a neural network model
How to build a Convolutional Neural Network that works well with images
How to use that model to make predictions on new images
How to turn those predictions into a submission to the competition